# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.InitiatePaymentRequest:
    properties:
      orderId:
        type: string
      paymentMethod:
        type: string
    type: object
...
  request.RefundRequest:
    properties:
      reason:
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      productId:
        type: string
      retryAfter:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
      success:
        type: boolean
    type: object
...
  response.ProductListResponse:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      products:
        items:
          $ref: '#/definitions/response.ProductResponse'
        type: array
      total:
        type: integer
    type: object
  response.ProductResponse:
    properties:
      category:
        type: string
      categoryId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      price:
        type: number
      productId:
        type: string
      status:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: CS362 Team
  description: Flash Sale E-commerce API with Rate Limiting — CS362 Final Project
  termsOfService: http://swagger.io/terms/
  title: CS362 E-commerce Order System API
  version: "1.0"
paths:
...
  /payments:
    post:
      consumes:
      - application/json
      description: Creates a new payment record for an existing order and processes
        it through the payment gateway
      parameters:
      - description: Payment request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.InitiatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Initiate a payment
      tags:
      - Payments
  /payments/{paymentId}/refund:
    post:
      consumes:
      - application/json
      description: Initiates a refund for a completed payment. Only COMPLETED payments
        can be refunded.
      parameters:
      - description: Payment ID
        in: path
        name: paymentId
        required: true
        type: string
      - description: Refund reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Payment not eligible for refund
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Refund a payment
      tags:
      - Payments
  /products:
    get:
      description: Returns a paginated list of available products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ProductListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List all products
      tags:
      - Products
  /products/{productId}:
    get:
      description: Retrieves a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ProductResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get product by ID
      tags:
      - Products
  /products/search:
    get:
      description: Returns products matching a search keyword in name or description
      parameters:
      - description: Search keyword
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.ProductListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Search products
      tags:
      - Products
securityDefinitions:
  CustomerID:
    description: Customer identifier passed as request header
    in: header
    name: X-Customer-ID
    type: apiKey
swagger: "2.0"